v2/requirements.txt
gradio==5.30.0
pymongo==4.13.0
python-dotenv==1.0.1
google-genai==0.2.0 